Add files via upload
The purpose of this project is to implement one of the types of GAN neural networks called DCGAN. The main difference between DCGAN and other such networks is the use of convolution layers in it, and the results are far better than other GAN networks. Various layers such as Conv2DTranspose, BatchNormalization, LeakyReLU and Dropout have been used to implement this network.